PMM-6091 Rename UpdateModal to ProgressModal
