+++ pmm-app/src/pmm-update/components/ProgressModal/ProgressModal.styles.ts
