+++ pmm-app/src/pmm-update/components/index.ts
@@ -4,5 +4,5 @@ export { LastCheck } from './LastCheck/LastCheck';
 export { UpdateButton } from './UpdateButton/UpdateButton';
 export { UpdateHeader } from './UpdateHeader/UpdateHeader';
 export { UpdateInfoBox } from './UpdateInfoBox/UpdateInfoBox';
-export { UpdateModal } from './UpdateModal/UpdateModal';
+export { ProgressModal } from './ProgressModal/ProgressModal';
 export { UpdateModalHeader } from './UpdateModalHeader/UpdateModalHeader';
